[FB] [PI-3478] Lenient cube arithmetic (#3774)
* initial cube arithmetic

* support in-place cube resolve

* fix non in-place broadcasting

* remove temporary resolve scenario test

* lenient/strict support for attributes dicts with numpy arrays

* lenient/strict treatment of scalar coordinates

* strict points/bounds matching

* lenient/strict prepare local dim/aux/scalar coordinates

* support extended broadcasting

* always raise exception on points/bounds mismatch

* ignore scalar points/bounds mismatches, lenient only

* remove todos

* tidy logger debugs

* qualify src/tgt cube references in debug

* Numpy rounding fix (#3758)

ensure rounding is numpy like (maintains type)

* avoid unittest.mock.sentinel copy issue

* fast load np.int32

* fix cube maths doctest

* fix iris.common.resolve logging configuration

* fix prepare points/bounds + extra metadata cml

* support mapping reversal based on free dims

* var_name fix for lenient equality

* add support for DimCoordMetadata

* fix circular flag + support CoordMetadata and DimCoordMetadata exchange

* fix circular issue for concatenate DimCoord->AuxCoord demotion

* fix concatenate._CubeSignature sorted

* minor tweaks

* keep lenient_client private in maths

* tidy maths

* tidy iris.analysis.maths.IFunc

* refactor IFunc test

* polish in-place support

* tidy metadata_resolve

+++ lib/iris/tests/integration/fast_load/test_fast_load.py
@@ -377,7 +377,8 @@ class MixinCallDetails:
         # Make an 'expected' from selected fields, with the expected attribute.
         expected = CubeList([flds[1], flds[3]]).merge()
         if not self.do_fast_loads:
-            expected[0].attributes["LBVC"] = 8
+            # This is actually a NumPy int32, so honour that here.
+            expected[0].attributes["LBVC"] = np.int32(8)
         else:
             expected[0].attributes["A_LBVC"] = [8, 8]
